Refactor from UByte to Byte

+++ app/src/main/kotlin/com/mlesniak/gameboy/debug/Conversions.kt
@@ -3,8 +3,8 @@ package com.mlesniak.gameboy.debug
 import kotlin.math.max
 import kotlin.math.min
 
-fun UInt.hex(padding: Int = 0) = "0x" + ("%X".format(this.toInt()).padStart(padding, '0'))
+fun Int.hex(padding: Int = 0) = "0x" + ("%X".format(this).padStart(padding, '0'))
 
-fun UByte.hex(padding: Int = 0) = "0x" + ("%X".format(this.toByte()).padStart(padding, '0'))
+fun Byte.hex(padding: Int = 0) = "0x" + ("%X".format(this).padStart(padding, '0'))
 
-fun UInt.clamp(minVal: UInt, maxVal: UInt = UInt.MAX_VALUE) = max(min(this, maxVal), minVal)
+fun Int.clamp(minVal: Int, maxVal: Int = Int.MAX_VALUE) = max(min(this, maxVal), minVal)
